Editable: Update state.cursorOffset on edit. Fixes #3037.
Otherwise it can become out of sync with the cursor thought value and cause a selection error on undo when the selection is being restored.

Depending on what is subscribed to state.cursorOffset, this commit could result in additional component renders. But the `thoughtChangeHandler` in the Editable component is already throttled, and the thought value itself is changing at the same time, so hopefully this does not create any additional renders in practice. A component would have to subscribe to state.cursorOffset but not the thought value for this to happen.